Combined ContactsFaults and OrientationPoints symbology into a single class
